updates to file manager

+++ desipipe/tests/test_task_manager.py
@@ -19,9 +19,24 @@ def test_app():
     print(app.run((1, 1), {}))
 
 
-def test_queue():
+def test_serialization():
 
-    spawn = True
+    from desipipe.task_manager import serialize_function, deserialize_function
+
+    def func(a, b, *args, c=4, **kwargs):
+        return a * b * c
+
+    name, code, dlocals = serialize_function(func)
+    func2 = deserialize_function(name, code, dlocals)
+    assert func2(1, 2, c=5) == 10.
+    name, code, dlocals = serialize_function(func2)
+    print(name, code, dlocals)
+    func2 = deserialize_function(name, code, dlocals)
+    assert func2(1, 2) == 8.
+
+
+def test_queue(spawn=True, run=False):
+
     queue = Queue('test', base_dir=base_dir, spawn=spawn)
     provider = None
     if os.getenv('NERSC_HOST', None):
@@ -29,13 +44,19 @@ def test_queue():
     tm = TaskManager(queue, environ=dict(), scheduler=dict(max_workers=2), provider=provider)
     tm2 = tm.clone(scheduler=dict(max_workers=1), provider=dict(provider='local'))
 
-    @tm.python_app
-    def fraction(size=10000):
+    def common1(size):
         import time
         import numpy as np
-        time.sleep(5)
+        time.sleep(2)
         x, y = np.random.uniform(-1, 1, size), np.random.uniform(-1, 1, size)
         return np.sum((x**2 + y**2) < 1.) * 1. / size
+
+    def common2(size, co=common1):
+        return co(size=size)
+
+    @tm.python_app
+    def fraction(size=10000, co=common2):
+        return co(size=size)
 
     @tm2.bash_app
     def echo(fractions):
@@ -61,6 +82,7 @@ def test_queue():
     avg2 = average2(fractions)
     assert average3(fractions) is None
     if spawn:
+        print(queue.summary())
         print(ech.out())
         assert avg2.result() == avg.result()
         print(avg.result(), time.time() - t0)
@@ -72,10 +94,19 @@ def test_queue():
     if spawn:
         for frac in fractions:
             assert fraction().result() == frac.result()  # the previous fraction result is used
+    elif run:
+        from desipipe.task_manager import TaskPickler, TaskUnpickler
+        task = queue.pop()
+        pkl = TaskPickler.dumps(task, reduce_app=False)
+        task = TaskUnpickler.loads(pkl)
+        print(task.kwargs)
+        task.run()
+        print(task.err)
 
-    for tid in queue.tasks(name='fraction', property='tid'):
-        del queue[tid]
-    queue.delete()
+    if spawn:
+        for tid in queue.tasks(name='fraction', property='tid'):
+            del queue[tid]
+        queue.delete()
 
 
 def test_cmdline():
@@ -92,17 +123,16 @@ def test_cmdline():
     subprocess.call(['desipipe', 'retry', '-q', queue, '--state', 'SUCCEEDED', '--spawn'])
 
 
-def test_file():
+def test_file(spawn=True):
 
     txt = 'hello world!'
 
     fm = FileManager()
-    fm.db.append(dict(description='added file', id='input', filetype='text', path=os.path.join(base_dir, 'hello_in_{i:d}.txt'), options={'i': range(10)}))
+    fm.append(dict(description='added file', id='input', filetype='text', path=os.path.join(base_dir, 'hello_in_{i:d}.txt'), options={'i': range(10)}))
     for fi in fm:
         fi.get().write(txt)
-    fm.db.append(fm.db[0].clone(id='output', path=os.path.join(base_dir, 'hello_out_{i:d}.txt')))
+    fm.append(fm[0].clone(id='output', path=os.path.join(base_dir, 'hello_out_{i:d}.txt')))
 
-    spawn = False
     queue = Queue('test2', base_dir=base_dir, spawn=spawn)
     provider = None
     if os.getenv('NERSC_HOST', None):
@@ -130,7 +160,8 @@ def test_file():
 
 if __name__ == '__main__':
 
+    #test_serialization()
     #test_app()
-    test_queue()
+    #test_queue(spawn=True)
     #test_cmdline()
-    test_file()
+    test_file(spawn=True)
